base layout jss cleanup
